modify

+++ DaeguDay/DaeguDay/src/main/java/com/rockies/DaeguDay/domain/Category.java
@@ -3,13 +3,13 @@ package com.rockies.DaeguDay.domain;
 import javax.persistence.Entity;
 
 public class Category {
-    private int number;
+    private Long number;
 
-    public int getNumber() {
+    public Long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(Long number) {
         this.number = number;
     }
 }
