+++ DaeguDay/DaeguDay/src/main/java/com/rockies/DaeguDay/domain/News.java
@@ -16,6 +16,6 @@ public class News {
     private String headline;
     @Id
     private String url;
-    private int category;
+    private long category;
 
 }
